fix: preserve source line limits after develop merge
