+++ crates/atm-http-runtime/src/router_support.rs
@@ -1,6 +1,8 @@
+use atm_core::LocalServiceRuntime;
 use atm_core::error::AtmError;
 use atm_core::protocol::{ResponseEnvelope, SendResponseEnvelope};
 use atm_core::send::{WarningEntry, WriteOutcome};
+use atm_core::types::{AgentName, TeamName};
 
 pub(super) fn append_warnings(outcome: &mut WriteOutcome, warnings: Vec<WarningEntry>) {
     match outcome {
@@ -24,4 +26,15 @@ pub(super) fn hook_warning(error: AtmError) -> WarningEntry {
         format!("message received successfully, but its receiver hook did not run: {error}"),
         Some("inspect the receiver hook endpoint or harness, then continue normally"),
     )
+}
+
+pub(super) fn validate_graft_receiver_member(
+    runtime: &LocalServiceRuntime,
+    team: &TeamName,
+    agent: &AgentName,
+) -> Result<(), AtmError> {
+    if runtime.load_roster_member(team, agent)?.is_none() {
+        return Err(AtmError::agent_not_found(agent.as_str(), team.as_str()));
+    }
+    Ok(())
 }
